+++ src/graph.cc
@@ -408,12 +408,14 @@ struct EdgeEnv : public Env {
 
 string EdgeEnv::LookupVariable(const string& var) {
   if (var == "in" || var == "in_newline") {
-    int explicit_deps_count = edge_->inputs_.size() - edge_->implicit_deps_ -
-      edge_->order_only_deps_;
+    int explicit_deps_count =
+        static_cast<int>(edge_->inputs_.size() - edge_->implicit_deps_ -
+                         edge_->order_only_deps_);
     return MakePathList(edge_->inputs_.data(), explicit_deps_count,
                         var == "in" ? ' ' : '\n');
   } else if (var == "out") {
-    int explicit_outs_count = edge_->outputs_.size() - edge_->implicit_outs_;
+    int explicit_outs_count =
+        static_cast<int>(edge_->outputs_.size() - edge_->implicit_outs_);
     return MakePathList(&edge_->outputs_[0], explicit_outs_count, ' ');
   }
 
@@ -706,7 +708,7 @@ bool ImplicitDepLoader::ProcessDepfileDeps(
     Edge* edge, std::vector<StringPiece>* depfile_ins, std::string* err) {
   // Preallocate space in edge->inputs_ to be filled in below.
   vector<Node*>::iterator implicit_dep =
-      PreallocateSpace(edge, depfile_ins->size());
+      PreallocateSpace(edge, static_cast<int>(depfile_ins->size()));
 
   // Add all its in-edges.
   for (std::vector<StringPiece>::iterator i = depfile_ins->begin();
